Switched to dropdown menu for Description titles.
Why?

Because then it will cause me to get into the habits of going through
the motions of describing certain things I should for every sprint and
card, such as "what the end looks like," "potential pitfalls," "things
that need to stay constant, should not change," "potential strategies
for implementing the card," which may help guide the ai programmer.

Etc.